Moves libraries into lib folder.
