+++ lib/assets_helper.py
