+++ viewer.py
@@ -15,8 +15,8 @@ import sh
 from settings import settings
 import html_templates
 from lib.utils import url_fails
-import db
-import assets_helper
+from lib import db
+from lib import assets_helper
 
 
 __author__ = "WireLoad Inc"
